MPA-9: Move biometric authentication related vars from UserDefaultsClient.
